gitautosync update 14

+++ mysum.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-def sum_numbers(*numbers: int) -> int:
+def sum_numbers(*numbers: float) -> float:
     total = 0
 
     for one_number in numbers:
